refactor(rag-knowledge): U1 完了 — format 関数を境界整形ロジック層に切り出し (#730)
Issue #730 の Unit 1 完了。format 関数を CLI / MCP 共有の境界整形ロジック層
（package 直下の formatting.py）に切り出し、`rag_knowledge.py` を縮小する。

- 新規: `src/rag/search/formatting.py`
  - `format_raw_search_results` / `_format_chunk_position`
- 新規: `src/rag/admin/formatting.py`
  - `format_document_response` / `format_file_size`
- 改訂: `src/rag/rag_knowledge.py`
  - format 関数の元定義を削除（dto / smart_chunk と同様に新配置を import）
  - LOC: 約 880 → 782
- 改訂: import 経路の更新
  - `src/rag/cli.py`: `format_document_response` / `format_raw_search_results` /
    `format_file_size` の import 元
  - `src/rag/server/tools/listing.py`: `format_file_size` の import 元
  - `tests/test_get_document.py` / `tests/test_list_recent.py` /
    `tests/test_rebuild_stats.py`: 同上

format 関数は CLI と MCP 両方から利用されるため（特に `format_file_size` は
CLI と server/tools/listing.py 両方で使用）、Q4=B「CLI / MCP 応答層に取り込み」
の解釈を「境界整形ロジックを Service から外す」と捉え、各 dto package 直下の
formatting.py に配置した。これは architecture spec §5 軸 2 の Port 配置原則
（package 内集約）と整合する。
